+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/AbstractModInitializer.java
@@ -13,18 +13,19 @@ import com.refinedmods.refinedstorage.common.api.RefinedStorageApi;
 import com.refinedmods.refinedstorage.common.api.RefinedStorageApiProxy;
 import com.refinedmods.refinedstorage.common.api.security.PlatformSecurityNetworkComponent;
 import com.refinedmods.refinedstorage.common.api.upgrade.AbstractUpgradeItem;
-import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterBlockEntity;
-import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterContainerMenu;
-import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterData;
 import com.refinedmods.refinedstorage.common.autocrafting.CraftingPatternState;
-import com.refinedmods.refinedstorage.common.autocrafting.patterngrid.PatternGridBlockEntity;
-import com.refinedmods.refinedstorage.common.autocrafting.patterngrid.PatternGridContainerMenu;
-import com.refinedmods.refinedstorage.common.autocrafting.patterngrid.PatternGridData;
 import com.refinedmods.refinedstorage.common.autocrafting.PatternItem;
 import com.refinedmods.refinedstorage.common.autocrafting.PatternState;
 import com.refinedmods.refinedstorage.common.autocrafting.ProcessingPatternState;
 import com.refinedmods.refinedstorage.common.autocrafting.SmithingTablePatternState;
 import com.refinedmods.refinedstorage.common.autocrafting.StonecutterPatternState;
+import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterBlockEntity;
+import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterContainerMenu;
+import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterData;
+import com.refinedmods.refinedstorage.common.autocrafting.autocraftermanager.AutocrafterManagerBlockEntity;
+import com.refinedmods.refinedstorage.common.autocrafting.patterngrid.PatternGridBlockEntity;
+import com.refinedmods.refinedstorage.common.autocrafting.patterngrid.PatternGridContainerMenu;
+import com.refinedmods.refinedstorage.common.autocrafting.patterngrid.PatternGridData;
 import com.refinedmods.refinedstorage.common.configurationcard.ConfigurationCardItem;
 import com.refinedmods.refinedstorage.common.configurationcard.ConfigurationCardState;
 import com.refinedmods.refinedstorage.common.constructordestructor.BlockBreakDestructorStrategyFactory;
@@ -343,6 +344,7 @@ public abstract class AbstractModInitializer {
         Blocks.INSTANCE.getRelay().registerBlocks(callback);
         Blocks.INSTANCE.setDiskInterface(blockEntityProviders.diskInterface()).registerBlocks(callback);
         Blocks.INSTANCE.getAutocrafter().registerBlocks(callback);
+        Blocks.INSTANCE.getAutocrafterManager().registerBlocks(callback);
     }
 
     protected final void registerItems(final RegistryCallback<Item> callback) {
@@ -366,6 +368,7 @@ public abstract class AbstractModInitializer {
         Blocks.INSTANCE.getRelay().registerItems(callback, Items.INSTANCE::addRelay);
         Blocks.INSTANCE.getDiskInterface().registerItems(callback, Items.INSTANCE::addDiskInterface);
         Blocks.INSTANCE.getAutocrafter().registerItems(callback, Items.INSTANCE::addAutocrafter);
+        Blocks.INSTANCE.getAutocrafterManager().registerItems(callback, Items.INSTANCE::addAutocrafterManager);
         registerStorageItems(callback);
         registerUpgrades(callback);
     }
@@ -667,6 +670,11 @@ public abstract class AbstractModInitializer {
         BlockEntities.INSTANCE.setAutocrafter(callback.register(
             ContentIds.AUTOCRAFTER,
             () -> typeFactory.create(AutocrafterBlockEntity::new, Blocks.INSTANCE.getAutocrafter().toArray())
+        ));
+        BlockEntities.INSTANCE.setAutocrafterManager(callback.register(
+            ContentIds.AUTOCRAFTER_MANAGER,
+            () -> typeFactory.create(AutocrafterManagerBlockEntity::new,
+                Blocks.INSTANCE.getAutocrafterManager().toArray())
         ));
     }