+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/Config.java
@@ -75,6 +75,8 @@ public interface Config {
 
     AutocrafterEntry getAutocrafter();
 
+    SimpleEnergyUsageEntry getAutocrafterManager();
+
     interface SimpleEnergyUsageEntry {
         long getEnergyUsage();
     }